requirements.txt
fastapi
uvicorn
sqlalchemy
pillow
moviepy